satellite/requirements.txt
# Base system
paho-mqtt==1.6.1
psutil
numpy
pandas

# Minimal anomaly models support
torch                  # per Chronos
chronos-forecasting    # modello Chronos
nixtla                 # TimeGPT client
pyod                   # Unsupervised (LOF, INNE, IForest)
